Fix some bugs
The camera after authorization can use the focus for the first time using.
The custom camera view is updated to fit with all kind of screen.
A toolbar is added in the main fragment to explain the purpose af the app with a simple title, a about button will be added soon.
The onBackPressed is disabled to avoid bad user experience navigating strangely between fragment end camera activity.
Switch relative layout with constraint layout